lesson-34: Add lesson-34 task employees-app GlobalExceptionHandler and create service interface
