+++ sprng-boot/lesson-34/employees-app/src/main/java/az/atl/academy/employees/app/model/constants/ExceptionConstants.java
@@ -1,0 +1,9 @@
+package az.atl.academy.employees.app.model.constants;
+
+public final class ExceptionConstants {
+    public static final String EMPLOYEE_NOT_FOUND_CODE = "EMPLOYEE_NOT_FOUND";
+    public static final String EMPLOYEE_NOT_FOUND_MESSAGE = "Employee with id %s not found";
+
+    public static final String DEPARTMENT_NOT_FOUND_CODE = "TUTORIAL_ALREADY_EXISTS";
+    public static final String DEPARTMENT_NOT_FOUND_MESSAGE = "Department with id %s doesn't exist";
+}
